If, else, goto not recursive
This modifies the code so that having goto doesn't cause the parse function to recurse. It changes the initial parse loop (line 143) from a for loop, to a while loop, this way you can modify the location where the interpreter is looking for code, allowing gotos to be non-recursive.

The if statement is quite messy, the first thing it does is split the line by spaces. This would be fine unless the user is comparing string literals with spaces, in which case the string literal gets split as well, there is a for loop later on that concatenates it (That for loop does not have to make sense, it works so it works.

The code then checks if the if statement compares variables, if it does, the variables will be replaced with the values of the variables. Finally, several if and elif statements check what type of if statement the user wants, and then sets evaluation variable based on that.

last evaluation is also set to globally track what the last if statement does, this is for the else statements later.

The else statement checks if the last statement was an if, it throws an error if not.

Then it checks the last evaluation variable set earlier, it true, it goes to the goto line.